\name{jedit.pager}
\Rdversion{1.1}
\alias{jedit.pager}
\title{
jedit pager
}
\description{
jedit pager. This is used by this package as the pager option
in order to display files in jedit
}
\usage{
jedit.pager(files, header, title, delete.file)
}
\arguments{
  \item{files}{files to display}
  \item{header}{unused}
  \item{title}{unused}
  \item{delete.file}{unused}
}
\value{
NULL
}
\author{
Romain Francois <francoisromain@free.fr>
}
\seealso{
\code{\link{options}}, specially the section about "pager"
}
\keyword{ manip }

