\name{orchestra-package}
\Rdversion{1.1}
\alias{orchestra-package}
\alias{orchestra}
\docType{package}
\title{
support for embedding R into jedit
}
\description{
support for embedding R into jedit
}
\details{
\tabular{ll}{
Package: \tab orchestra\cr
Type: \tab Package\cr
Version: \tab 0.0-2\cr
Date: \tab 2009-08-06\cr
License: \tab GPL (>=3)\cr
LazyLoad: \tab yes\cr
}
}
\author{
Romain Francois <francoisromain@free.fr>
Bernd Bischl <bernd_bischl@gmx.net>

Maintainer: Romain Francois <francoisromain@free.fr>
}
\references{
	jedit. Programmer's text editor. \url{http://www.jedit.org/}
}
\keyword{ package }

