\name{textArea}
\Rdversion{1.1}
\alias{textArea}
\docType{data}
\title{
binding to the current jedit text area
}
\description{
this object is an active binding to the current jedit text area
}
\usage{textArea}
\format{
The binding resolves to a java reference to the currently active text area
}
\details{
The textArea object that is 
resolved is an instance of the 
\url{http://www.jedit.org/api/index.html?org/gjt/sp/jedit/textarea/JEditTextArea.html}
class
}
\keyword{datasets}

