\name{view}
\Rdversion{1.1}
\alias{view}
\docType{data}
\title{
binding to the current jedit view
}
\description{
this object is an active binding to the current jedit view
}
\usage{view}
\format{
The binding resolves to a java reference to the currently active view
}
\details{
The view objects that is 
resolved is an instance of the 
\url{http://www.jedit.org/api/index.html?org/gjt/sp/jedit/View.html}
class
}
\keyword{datasets}

